+++ src/build_scripts/copy_dir.py
@@ -3,7 +3,6 @@
 # dependencies = []
 # ///
 
-# ruff: noqa: INP001
 import logging
 import shutil
 import sys