+++ src/reactpy/widgets.py
@@ -2,12 +2,11 @@ from __future__ import annotations
 
 from base64 import b64encode
 from collections.abc import Sequence
-from typing import TYPE_CHECKING, Any, Callable, Protocol, TypeVar
+from typing import Any, Callable, Protocol, TypeVar
 
 import reactpy
 from reactpy._html import html
-from reactpy._warnings import warn
-from reactpy.types import ComponentConstructor, VdomAttributes, VdomDict
+from reactpy.types import VdomAttributes, VdomDict
 
 
 def image(
@@ -22,11 +21,7 @@ def image(
     if format == "svg":
         format = "svg+xml"  # noqa: A001
 
-    if isinstance(value, str):
-        bytes_value = value.encode()
-    else:
-        bytes_value = value
-
+    bytes_value = value.encode() if isinstance(value, str) else value
     base64_value = b64encode(bytes_value).decode()
     src = f"data:image/{format};base64,{base64_value}"
 
@@ -83,20 +78,3 @@ _CastTo_co = TypeVar("_CastTo_co", covariant=True)
 
 class _CastFunc(Protocol[_CastTo_co]):
     def __call__(self, value: str) -> _CastTo_co: ...
-
-
-if TYPE_CHECKING:
-    from reactpy.testing.backend import _MountFunc
-
-
-def hotswap(
-    update_on_change: bool = False,
-) -> tuple[_MountFunc, ComponentConstructor]:  # nocov
-    warn(
-        "The 'hotswap' function is deprecated and will be removed in a future release",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    from reactpy.testing.backend import _hotswap
-
-    return _hotswap(update_on_change)